Melt SupportsAdmissionControl into stream reader
