+++ python/pyspark/sql/datasource.py
@@ -714,25 +714,55 @@ class DataSourceStreamReader(ABC):
             messageParameters={"feature": "initialOffset"},
         )
 
-    def latestOffset(self) -> dict:
+    from pyspark.sql.streaming.datasource import ReadAllAvailable, ReadLimit
+
+    # FIXME: Will this work without abstractmethod marker?
+    # @abstractmethod
+    def latestOffset(self, start: dict, readLimit: ReadLimit) -> dict:
         """
-        Returns the most recent offset available.
+        FIXME: docstring needed
 
-        Returns
-        -------
-        dict
-            A dict or recursive dict whose key and value are primitive types, which includes
-            Integer, String and Boolean.
-
-        Examples
-        --------
-        >>> def latestOffset(self):
-        ...     return {"parititon-1": {"index": 3, "closed": True}, "partition-2": {"index": 5}}
+        /**
+        * Returns the most recent offset available given a read limit. The start offset can be used
+        * to figure out how much new data should be read given the limit. Users should implement this
+        * method instead of latestOffset for a MicroBatchStream or getOffset for Source.
+        * <p>
+        * When this method is called on a `Source`, the source can return `null` if there is no
+        * data to process. In addition, for the very first micro-batch, the `startOffset` will be
+        * null as well.
+        * <p>
+        * When this method is called on a MicroBatchStream, the `startOffset` will be `initialOffset`
+        * for the very first micro-batch. The source can return `null` if there is no data to process.
+        */
         """
         raise PySparkNotImplementedError(
             errorClass="NOT_IMPLEMENTED",
             messageParameters={"feature": "latestOffset"},
         )
+
+    def getDefaultReadLimit(self) -> ReadLimit:
+        """
+          FIXME: docstring needed
+
+        /**
+         * Returns the read limits potentially passed to the data source through options when creating
+         * the data source.
+         */
+        """
+        return ReadAllAvailable()
+
+    def reportLatestOffset(self) -> Optional[dict]:
+        """
+          FIXME: docstring needed
+
+        /**
+         * Returns the most recent offset available.
+         * <p>
+         * The source can return `null`, if there is no data to process or the source does not support
+         * to this method.
+         */
+        """
+        return None
 
     def partitions(self, start: dict, end: dict) -> Sequence[InputPartition]:
         """
